+++ tests/unit/test_solvers/test_base_solver.py
@@ -50,6 +50,17 @@ class TestBaseSolver(unittest.TestCase):
         ):
             solver.solve(model, np.array([1, 2, 3, 2]))
 
+    def test_block_symbolic_inputs(self):
+        solver = pybamm.BaseSolver(rtol=1e-2, atol=1e-4)
+        model = pybamm.BaseModel()
+        a = pybamm.Scalar(0)
+        p = pybamm.InputParameter("p")
+        model.rhs = {a: a * p}
+        with self.assertRaisesRegex(
+            pybamm.SolverError, "Only CasadiAlgebraicSolver can have symbolic inputs"
+        ):
+            solver.solve(model, np.array([1, 2, 3]))
+
     def test_ode_solver_fail_with_dae(self):
         model = pybamm.BaseModel()
         a = pybamm.Scalar(1)